Validation interface and AccountDtoValidator is done, one test passes.

+++ src/main/java/com/sda/dmpcars/service/AccountService.java
@@ -5,7 +5,10 @@ import com.sda.dmpcars.dto.AccountDto;
 import com.sda.dmpcars.model.Account;
 import com.sda.dmpcars.model.AccountDetail;
 import com.sda.dmpcars.model.AccountType;
+import com.sda.dmpcars.validator.AccountDtoValidator;
+import com.sda.dmpcars.validator.Validator;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -15,31 +18,34 @@ import java.util.List;
 public class AccountService {
 
     private final AccountDao accountDao;
+    private final Validator<AccountDto> validator = new AccountDtoValidator();
 
     @Autowired
     public AccountService(AccountDao accountDao) {
         this.accountDao = accountDao;
     }
 
-    public void addAccount(AccountDto accountDto){
-
-        Account account = Account.builder().id(accountDto.getId()).login(accountDto.getLogin())
-                .password(accountDto.getPassword()).accountDetail(
-                        AccountDetail.builder().firstName(accountDto.getFirstName())
-                                .lastName(accountDto.getLastName()).email(accountDto.getEmail())
-                                .phoneNumber(accountDto.getPhoneNumber()).yearOfBirth(accountDto.getYearOfBirth())
-                                .build()
-                ).accountType(
-                        AccountType.builder().role("USER").build()
-                ).build();
-
-        accountDao.save(account);
+    public boolean addAccount(AccountDto accountDto) {
+        if (validator.validate(accountDto)) {
+            Account account = Account.builder().id(accountDto.getId()).login(accountDto.getLogin())
+                    .password(accountDto.getPassword()).accountDetail(
+                            AccountDetail.builder().firstName(accountDto.getFirstName())
+                                    .lastName(accountDto.getLastName()).email(accountDto.getEmail())
+                                    .phoneNumber(accountDto.getPhoneNumber()).yearOfBirth(accountDto.getYearOfBirth())
+                                    .build()
+                    ).accountType(
+                            AccountType.builder().role("USER").build()
+                    ).build();
+            accountDao.save(account);
+            return true;
+        }
+        return false;
     }
 
-    public AccountDto getAccountById(Integer id){
+    public AccountDto getAccountById(Integer id) {
 
         Account account = accountDao.findById(id).orElse(null);
-        if(account != null){
+        if (account != null) {
             return AccountDto.builder().id(account.getId()).login(account.getLogin())
                     .password(account.getPassword()).firstName(account.getAccountDetail().getFirstName())
                     .lastName(account.getAccountDetail().getLastName()).email(account.getAccountDetail().getEmail())
@@ -51,7 +57,7 @@ public class AccountService {
         return new AccountDto();
     }
 
-    public List<AccountDto> getAllAccounts(){
+    public List<AccountDto> getAllAccounts() {
         List<AccountDto> result = new ArrayList<>();
 
         accountDao.findAll().forEach(account -> AccountDto.builder().id(account.getId()).login(account.getLogin())
@@ -64,30 +70,36 @@ public class AccountService {
         return result;
     }
 
-    public void deleteAccount(AccountDto accountDto){
-        Account account = Account.builder().id(accountDto.getId()).login(accountDto.getLogin()).build();
-        accountDao.delete(account);
+    public void deleteAccount(AccountDto accountDto) {
+        if (validator.validate(accountDto)) {
+            Account account = Account.builder().id(accountDto.getId()).login(accountDto.getLogin()).build();
+            accountDao.delete(account);
+        }
     }
 
-    public void deleteAllAccounts(){
+    public void deleteAllAccounts() {
         accountDao.deleteAll();
     }
 
-    public AccountDto updateAccount(AccountDto accountDto){
-        Account account = Account.builder().id(accountDto.getId()).login(accountDto.getLogin())
-                .password(accountDto.getPassword()).accountDetail(
-                        AccountDetail.builder().firstName(accountDto.getFirstName())
-                                .lastName(accountDto.getLastName()).email(accountDto.getEmail())
-                                .phoneNumber(accountDto.getPhoneNumber()).yearOfBirth(accountDto.getYearOfBirth())
-                                .build()).build();
+    public AccountDto updateAccount(AccountDto accountDto) {
+        if (validator.validate(accountDto)) {
+            Account account = Account.builder().id(accountDto.getId()).login(accountDto.getLogin())
+                    .password(accountDto.getPassword()).accountDetail(
+                            AccountDetail.builder().firstName(accountDto.getFirstName())
+                                    .lastName(accountDto.getLastName()).email(accountDto.getEmail())
+                                    .phoneNumber(accountDto.getPhoneNumber()).yearOfBirth(accountDto.getYearOfBirth())
+                                    .build()).build();
 
-        account = accountDao.save(account);
+            account = accountDao.save(account);
 
-        return AccountDto.builder().id(account.getId()).login(account.getLogin())
-                .password(account.getPassword()).firstName(account.getAccountDetail().getFirstName())
-                .lastName(account.getAccountDetail().getLastName()).email(account.getAccountDetail().getEmail())
-                .phoneNumber(account.getAccountDetail().getPhoneNumber())
-                .yearOfBirth(account.getAccountDetail().getYearOfBirth())
-                .build();
+            return AccountDto.builder().id(account.getId()).login(account.getLogin())
+                    .password(account.getPassword()).firstName(account.getAccountDetail().getFirstName())
+                    .lastName(account.getAccountDetail().getLastName()).email(account.getAccountDetail().getEmail())
+                    .phoneNumber(account.getAccountDetail().getPhoneNumber())
+                    .yearOfBirth(account.getAccountDetail().getYearOfBirth())
+                    .build();
+        }
+        return new AccountDto();
     }
 }
+
